way-back-home/solutions/level_2/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "google-adk>=1.18.0",
    "google-cloud-aiplatform>=1.133.0",
    "google-cloud-spanner>=3.61.0",
    "google-cloud-storage>=3.7.0",
    "httpx>=0.28.1",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
    "websockets>=15.0.1",
]
